Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 // Your code here...
 #include<stdio.h>
 int main(){
-    int n;
-    scanf("%d",&n);
-    printf("You entered:,%d",n);
+    float n;
+    scanf("%f",&n);
+    printf("You entered:,%f",n);
 }
